+++ src/omnisharp/OmnisharpManager.ts
@@ -3,26 +3,20 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import * as util from '../common';
 import * as path from 'path';
 import * as semver from 'semver';
-import * as vscode from 'vscode';
-import { Logger } from '../logger';
+import * as util from '../common';
 import { OmnisharpDownloader } from './OmnisharpDownloader';
-import TelemetryReporter from 'vscode-extension-telemetry';
 import { PlatformInformation } from '../platform';
 
 export class OmnisharpManager {
     public constructor(
-        private channel: vscode.OutputChannel,
-        private logger: Logger,
-        private packageJSON: any,
-        private reporter?: TelemetryReporter) {
+        private downloader: OmnisharpDownloader,
+        private platformInfo: PlatformInformation) {
     }
 
-    public async GetOmnisharpPath(omnisharpPath: string, useMono: boolean, serverUrl: string, latestVersionFileServerPath: string, installPath: string, extensionPath: string, platformInfo: PlatformInformation): Promise<string> {
+    public async GetOmnisharpPath(omnisharpPath: string, useMono: boolean, serverUrl: string, latestVersionFileServerPath: string, installPath: string, extensionPath: string): Promise<string> {
         // Looks at the options path, installs the dependencies and returns the path to be loaded by the omnisharp server
-        let downloader = new OmnisharpDownloader(this.channel, this.logger, this.packageJSON, platformInfo, this.reporter);
         if (path.isAbsolute(omnisharpPath)) {
             if (await util.fileExists(omnisharpPath)) {
                 return omnisharpPath;
@@ -32,22 +26,22 @@ export class OmnisharpManager {
             }
         }
         else if (omnisharpPath == "latest") {
-            return await this.InstallLatestAndReturnLaunchPath(downloader, useMono, serverUrl, latestVersionFileServerPath, installPath, extensionPath, platformInfo);
+            return await this.InstallLatestAndReturnLaunchPath(useMono, serverUrl, latestVersionFileServerPath, installPath, extensionPath);
         }
 
         //If the path is neither a valid path on disk not the string "latest", treat it as a version 
-        return await this.InstallVersionAndReturnLaunchPath(downloader, omnisharpPath, useMono, serverUrl, installPath, extensionPath, platformInfo);
+        return await this.InstallVersionAndReturnLaunchPath(omnisharpPath, useMono, serverUrl, installPath, extensionPath);
     }
 
-    private async InstallLatestAndReturnLaunchPath(downloader: OmnisharpDownloader, useMono: boolean, serverUrl: string, latestVersionFileServerPath: string, installPath: string, extensionPath: string, platformInfo: PlatformInformation) {
-        let version = await downloader.GetLatestVersion(serverUrl, latestVersionFileServerPath);
-        return await this.InstallVersionAndReturnLaunchPath(downloader, version, useMono, serverUrl, installPath, extensionPath, platformInfo);
+    private async InstallLatestAndReturnLaunchPath(useMono: boolean, serverUrl: string, latestVersionFileServerPath: string, installPath: string, extensionPath: string) {
+        let version = await this.downloader.GetLatestVersion(serverUrl, latestVersionFileServerPath);
+        return await this.InstallVersionAndReturnLaunchPath(version, useMono, serverUrl, installPath, extensionPath);
     }
 
-    private async InstallVersionAndReturnLaunchPath(downloader: OmnisharpDownloader, version: string, useMono: boolean, serverUrl: string, installPath: string, extensionPath: string, platformInfo: PlatformInformation) {
+    private async InstallVersionAndReturnLaunchPath(version: string, useMono: boolean, serverUrl: string, installPath: string, extensionPath: string) {
         if (semver.valid(version)) {
-            await downloader.DownloadAndInstallOmnisharp(version, serverUrl, installPath);
-            return GetLaunchPathForVersion(platformInfo, version, installPath, extensionPath, useMono);
+            await this.downloader.DownloadAndInstallOmnisharp(version, serverUrl, installPath);
+            return GetLaunchPathForVersion(this.platformInfo, version, installPath, extensionPath, useMono);
         }
         else {
             throw new Error(`Invalid omnisharp version - ${version}`);