+++ test/featureTests/OmnisharpDownloader.test.ts
@@ -3,7 +3,6 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import * as vscode from 'vscode';
 import * as path from 'path';
 import * as util from '../../src/common';
 import { should } from 'chai';
@@ -11,6 +10,7 @@ import { Logger } from '../../src/logger';
 import { OmnisharpDownloader } from '../../src/omnisharp/OmnisharpDownloader';
 import { rimraf } from 'async-file';
 import { PlatformInformation } from '../../src/platform';
+import { EventStream } from '../../src/EventStream';
 
 const tmp = require('tmp');
 const chai = require("chai");
@@ -20,7 +20,9 @@ let expect = chai.expect;
 suite("DownloadAndInstallExperimentalVersion : Gets the version packages, downloads and installs them", () => {
     let tmpDir = null;
     const version = "1.2.3";
-    const downloader = GetTestOmnisharpDownloader();
+    const platformInfo = new PlatformInformation("win32", "x86");
+    const eventStream = new EventStream();
+    const downloader = GetTestOmnisharpDownloader(eventStream, platformInfo);
     const serverUrl = "https://roslynomnisharp.blob.core.windows.net";
     const installPath = ".omnisharp/experimental/";
 
@@ -50,10 +52,8 @@ suite("DownloadAndInstallExperimentalVersion : Gets the version packages, downlo
     });
 });
 
-function GetTestOmnisharpDownloader() {
-    let channel = vscode.window.createOutputChannel('Experiment Channel');
-    let logger = new Logger(text => channel.append(text));
-    return new OmnisharpDownloader(channel, logger, GetTestPackageJSON(), new PlatformInformation("win32", "x86"), null);
+export function GetTestOmnisharpDownloader(sink: EventStream, platformInfo: PlatformInformation): OmnisharpDownloader{
+    return new OmnisharpDownloader(sink, GetTestPackageJSON(), platformInfo);
 }
 
 //Since we need only the runtime dependencies of packageJSON for the downloader create a testPackageJSON