+++ test/featureTests/index.ts
@@ -15,7 +15,9 @@
 // to report the results back to the caller. When the tests are finished, return
 // a possible error to the callback or null if none.
 
-let testRunner = require('vscode/lib/testrunner');
+import coverageWritingTestRunner from "../coverageWritingTestRunner";
+
+let testRunner = new coverageWritingTestRunner(require('vscode/lib/testrunner'));
 
 // You can directly control Mocha options by uncommenting the following lines
 // See https://github.com/mochajs/mocha/wiki/Using-mocha-programmatically#set-options for more info